chore: ui adjustments in swap/send/onramp flows
