+++ packages/appkit/src/views/w3m-swap-view/components/select-token-modal/styles.ts
@@ -8,7 +8,9 @@ export default StyleSheet.create({
     borderBottomLeftRadius: 0,
     borderBottomRightRadius: 0,
     overflow: 'hidden',
-    marginTop: 80
+    margin: 0,
+    flex: 1,
+    justifyContent: 'flex-end'
   },
   title: {
     paddingTop: Spacing['2xs']
@@ -29,7 +31,7 @@ export default StyleSheet.create({
     marginRight: Spacing.s
   },
   suggestedSeparator: {
-    marginVertical: Spacing.s
+    marginTop: Spacing.s
   },
   iconPlaceholder: {
     height: 32,
