+++ packages/appkit/src/views/w3m-swap-view/styles.ts
@@ -20,11 +20,7 @@ export default StyleSheet.create({
     marginTop: Spacing.xs,
     width: '100%'
   },
-  modalContent: {
-    margin: 0,
-    flex: 1,
-    justifyContent: 'flex-end'
-  },
+
   withKeyboard: {
     height: '100%'
   }
