+++ packages/appkit/src/views/w3m-wallet-send-view/styles.ts
@@ -19,6 +19,6 @@ export default StyleSheet.create({
     width: '100%'
   },
   withKeyboard: {
-    height: '100%'
+    height: '80%'
   }
 });
